fix open
